tools prometheus: allow scraping by ip address
This used to work because everything had a short alias with eqiad.wmflabs
on the end. It doesn't work anymore to use the hostname and port unless
something is in the tools project. Since we are monitoring things outside
the tools project, this allows you to specify using ip addresses instead.

metricsinfra scrapes the openstack api directly, but this doesn't at this
time.

Tested locally on tools-prometheus-03

Change-Id: I0389e837c712758138065134572a4c978649d8f8