gerrit: listen on all address with iptables rule
The Gerrit sshd service on port 29418 has been made to listen on
explicit address.  It was configured with:

  [sshd]
  listenAddress = gerrit.wikimedia.org
  listenAddress = [v6 service IP]:29418

When running `gerrit init`, the hostname is resolved to the IPv6 and
written back to gerrit.config. Gerrit then tries to listen twice on the
same IP, bails out on the second since the address is already in use and
Gerrit does not start. See T278990#7221954

In f9d4922e1 I have changed it to set an IPv4 explicitly, but `gerrit
init` somehow set it back to the IPv6. Maybe cause it does a reverse DNS
lookup and then a resolution.

The code in Gerrit is from 2009, at which time IPv6 was not yet in wide
spread use even at Google. It was convenient to just set the FQDN which
was solely resolving to IPv4.  Nowadays dual stack is way more frequent
and the handy DNS resolution is a bit fragile (ideally when given a
hostname Gerrit should resolve both v4 and v6 and listen on both).

Instead of trying to fix java code, remove `sshd.listenAddress` and rely
on Gerrit default to listen on addresses (`*:29418`).  That also solves
gerrit.config being altered when running `gerrit init`.

The intent to listen on explicit address was to prevent Gerrit from
listening on the host IP in addition of the service IP. Add firewall
rules to only allow traffic that are at the destination of the service
IPs.

Bug: T287122
Change-Id: I1b1e72473a83912e26a43070a3cd85c6c061ce11