wikidough: add motd script to indicate logging of root commands
As per Id438985ffe720dc630f0e43eed8bda4a47c9196c, for security reasons,
we will be logging all root commands on the Wikidough hosts with auditd.
Given that this is the first (and the only?) usecase for such logging,
we should indicate to users in the motd that all root commands on this
host will be logged and may be audited.

Change-Id: I066e94c28e4f38cc4eb885cc67601b658c5a9063