cloud dns: tidy up the labs-ip-alias-dump script
This is mostly stylistic changes. I haven't added a retry yet because
this script doesn't seem to fail often (from a scan of the logs), and
we are using different libraries now that would be interesting to
compare the performance of. Previously, we used novaclient to get info
from neutron, which regularly generates token errors for some reason.

One quirk is that changes (corrects?) the behavior when --check-changes-only
is not passed. There is no reason to return exit-status of 1 when this
succeeds in updating the DNS config! In fact, we may have been paged for that.

openstacksdk switches to neutron in the backend directly, so this is
more direct. This may resolve some of the problems in the original ticket
while retaining some of the better practices from I6291ea0b1e62560b7dda80ed9e2

Bug: T285537
Change-Id: I6a3c7b77459bacac7190cc1dfb5f68cbd6134988