wmflib::dir::mkdir_p: Use ensure_resource for full directory
Currently we create the directories and then apply the permissions to the
top level directory.  however this can cause an issues if two different
bits if could try and create the same directory.  This PS updates the
code so we first create the parent directories with no parameters and
then create the actual directory using the passed parameters

Hosts: cumin1001.eqiad.wmnet,install1003.wikimedia.org
Change-Id: Ic6d6f1a1c7bb65ed89a1d8dd40ea8a8642e46d9b