icinga: Add --services flag to icinga-status
By default, emit all services in states WARNING, CRITICAL, or UNKNOWN in
the json output. But if a regex is passed, instead emit all services
with matching names, regardless of their status. Targeted hosts with no
matching services are included in the output, with an empty service
list.

--services has no effect unless --json or --pretty-print is passed, and
no effect on the exit status (which is still 0 if all services are OK,
else 1).

Bug: T285803
Change-Id: I1fa3a46415b9f89b47e77863d9ce66812c6dd690