miscweb: Add CSP headers for query builder
This must be tested somewhere else before getting deployed to
production.

We can potentially extend this to all of query.wikidata.org gui but it
might break things, so one thing at a time.

Not adding CORS headers since this is a static serving website with no
write requests so I don't see the value in it.

Bug: T285761
Change-Id: I3e3ed4039ab48a70e19502bc0709dcdb67b15da5