trafficserver: limit mw on k8s to group0/test wikis
As the version of the wikis running on k8s is not always the same that
is  running in production, the concern of potential MW cache poisoning
has been raised, even if we total about 500 requests/day to this endpoint.

Given the potential for abuse by a third party, we're limiting
temporarily the access to just the test wikis and mediawiki.org, which
is part of the group0 wikis anyways.

Change-Id: I44db0653aa1720d08225311647e73662c8116b2c