backup: Move backup-related hosts from misc to new backup cluster
Move the following hosts:

* dbprov* from mysql to backup cluster
* ms-backup* from misc to backup cluster
* backup* from misc to backup cluster

db[12]* hosts, even if they are used just as a backup source or
backup metadata should stay in the mysql cluster, as in my mind
"being a database" is stronger than "it is being used for backups".

I think it makes sense to move dbprovs as while mysql is installed
there (for backup processing), it doesn't have a continously running
db.

Fix es and pc hosts, which were incorrectly appearing in the misc
cluster.

Bug: T276442
Change-Id: I46e5ddb1eb29105b6b99d7880da94e396d8d9ccc