site/conftool: convert mw1434-mw1436 from API to app servers
After renewed efforts to optimize balancing of appserver types
across rows, it became apparent that we have over 40% of appservers
in row D while we have 6 or 7 more API servers than appservers
in rows B and C and the grand total is 54 vs 62.

Therefore these 3 servers from C3 are being converted to the other
role.

It will turn the totals into 57 vs 59 and reduce the percentage
of appservers in row D vs the other rows.

Bug: T279309
Change-Id: I95c7e9bcc5333ff804a15ea925ce82270d5d4174