profile::kubernetes::deployment_server: add automation for mw on k8s
Currently, updating the image version for mediawiki on kubernetes
follows the standard gitops procedure we follow for every other service,
so we need to post and merge a patch to deployment-charts. This is
inconvenient given we already have a deployment procedure for mediawiki
outside of kubernetes.

As a very first solution, a simple script is introduced that does the
following:
* Check the latest released tag for the mediawiki-multiversion image
* update a special values file if a newer image is found
* if the file is updated, perform helmfile deploys to eqiad and codfw

This is of course an oversimplified model of release, and is not
expected to persist for long. It is, however, our way out of not being
able to allow browsing freely mwdebug on kubernetes for external clients.

Bug: T287570
Change-Id: Ie877156fbd62c9cf21f0a663b52f640fd9fd7fef