Update hive to log4j version2 configuration files
The current bigtop::hive module ships log4j configuration files that are
only applicable to log4j version1, whereas the version of hive that we
are running attempts to load version 2 files. The result is that we get
default logging parameters and cannot effect changes to the
configuration.

This commit adds new templates to the bigtop::hive module, based on the
templates shipped with hive. A new integer parameter is added to
profile::bigtop::client that enables the version of log4j to be
specified. The reason that the parameter is added to the client profile
is because this profile is included on all hive servers and clients.

When version 2 is specified the version 1 configuration files are
deleted from the host.

All hive clients and servers in the test cluster have version 2 defined.

Bug: T279304
Change-Id: I1d1d73efba0aec12503a04a821a2f1efc44ecb87