cloud-vps cloud-init: don't mask the puppet service
If we want cloud-init to configure puppet then it insists
on starting the puppet agent.  We don't really want
the puppet agent running; previously we masked it out
but that causes cloud-init to report failure to systemd;
now that we have wait_for_puppet() it might work to just
let the puppet agent complete a puppet run if it insists.

This means that firstboot will sometimes do three puppet runs
and sometimes two; I'm counting on idempotence to ensure
that this doesn't actually matter.

We waita good long time for the first puppet run to complete
if the agent started one; this should allow for complex
setups like toolforge nodes.

Bug: T287309
Change-Id: Ief51d591a45f67dabb021ddf49bf207dcc0d2ddc