bacula: Increase ES backups retention to 90 days
Database content (ES) backups retention was reduced to 60 days
when it had disk space issues on the old servers. Now that it
has been migrated to larger servers, increase the retention back
to the ideal time (90 days). Currently there is a 40% disk usage
and there is even more unallocated space available- so we could
use the same retention as in the database metadata backups.

This issue was spotted thanks to the new grafana backup cluster
graphs.

This requires followup deploy to apply the pool and volume
changes on the bacula director.

Bug: T282249
Change-Id: Ifec7fa1328c3b875e4c7332ee74815b15285eba5