bacula: Adjust number of max volumes for database backups
Due to the recent server migration and fs expansion, they
would be limited by the number of volumes, limiting the retention
if close the the max limit.

Allow content backups to take 100 TB and metadata backup 50TB of
disk space if necessary to make sure they have the right retention
(90 days), with some margin for emergencies.

Bug: T282249
Change-Id: I3e41f9ac3132025bf36aecff58fabb592470a0f9