Fix presto services in the test cluster
The previous change to the presto configuration in the test cluster
contained two errors and uncovered an issue with a required defined
type. This commit addresses the two errors in configuration, so that
presto can start on the test cluster again.

1) We failed to configure the test presto server to generate
   new certificates, so this contined to use the puppet certificates.

2) The option http.server.authentication.krb5.principal-hostname was
   added to the coordinator's configuration file, but this is not valid
   for the version of presto that we are using and thus caused an error.

The file /etc/presto/ssl/server.p12 also needs to be deleted manually
when applying this patch to the servers, because the current code will
not overwrite a valid file, even if the contents should be changed.

Bug: T273642
Change-Id: I4b5cd5b8ccabff5a548006b20647a96dbae7a719