Improve creation of pkcs12 file by checking contents
The existing sslcert::x509_to_pkcs12 function had a problem such that if
a valid output file already existed it would not everwrite it, even if
the contents should be changed according to the function's input
parameters.

This commit expands upon the 'unless' condition so that it checks the
certificate that is contained within the output file against that
specified in the input file. If the certificate is different or of the
output file does not exist, then the output file is generated. If it
matches, then creation of the file is skipped.

We do not check the contents of the private key, since we infer that if
the private key has changed, the certificate must have changed too.

Bug: T287869
Change-Id: Ia8269695225c1b7019ea72559b0ffeeedd6251a5