Set hive default log4j version to 2
We have made a change to the hive module in order to support log4j
versions 1 and 2, whilst we verified the version 2 logging on the test
cluster. Since that works as we expect, we are now changing the default
to version 2 in order to roll out this change to production.

Once all hive services and client sessions are restarted/completed with
this change, we will make another change to deprecate support for the
old version 1 support.

Bug: T279304
Change-Id: I38a671252ebd719d49ba59a0e72b408b6f4f8032