postgresql::user: split HBA configuration into a different define
Currently we tightly couple HBA configuration and user creation in
postgres configuration. This has been fine for the most part because
the main users of postgres connect to localhost (allowed by our
defaults in config), or they connect to the master in the cluster
(which is permitted by the existing setup).

However, currently when we configure a user and a set of network
access rules, the network access rules are applied *only* to the
masters. Replicas will have the user account replicated to them as
normal, but this user will _not_ have the network rules applied at the
same time.

In order to keep the current class configuration more or less the same
and not have to manage relationships between masters and replicas,
move HBA configuration to its own define. User creation is unaffected,
but for services like the maps cluster, HBA rules can now be defined
independent of user creation within the postgresql database
itself. This allows access to be granted independently on replicas by
using the hba define explicitly.

Bug: T283159
Change-Id: Iec833ade84b3d7333a57decf744239d59543adc3