Redirect jupyter notebook logs to logstash via kafka
We have to date been saving users' jupyter notebook logs to the local
disks, where they are not very accessible and can become large when
errors occur. This commit changes the configuration such that logs are
no longer saved to the local disks, but are directed to Logstash.

We add a suitable SyslogIdentifier to the systemd service unit and an
entry to the rsyslog lookup table in order to achieve this.

Bug: T287339
Change-Id: I16821d8a1b6cf6117efb9f2009db77521d8e3717