mediabackups: Switch TLS certificates to PKI rather than puppet
We reused the puppet certificates as a way to quickly test the
TLS functionality on media storage hosts.

While this worked -including TLS- for regular operations, Sadly,
those certs are unfit for production usage, as they use,
deprecately, the CN rather than SAN for host names. Switch to
using pki profile with the discovery intermediate CA, as we
think that will fix the tls error when scraping metrics from
prometheus.

Allow for undef TLS paths, mostly for puppet-compiler case.

Bug: T222113
Change-Id: I9acf974af669c9cf1ae212ed81500ac688ec086e