mediawiki: Migrate wikidatawiki dispatch crons to three systemd timers
Currently, using a complex system, a cron is being triggered which would
run for some certain period of time and then dies. This enables wikidata
to dispatch using concurrent processes (timeout time diveded by the time
interval of a new process would give number of concurrent processes)

This wouldn't work in systemd timers, it doesn't allow more than one
process being ran at the same time so it just checks that it's already
running and happily ignores the calendar trigger. To solve this, we
simply just introduce multiple systemd timers.

The whole this is a mess and should be replaced with jobs which is
tracked in T48643

Bug: T288175
Change-Id: I176c69454ec6a6996674cad3cdb8e673b15d271f