Create the druid user and group before installing druid-common
We found that when deploying a new driud server, a user account was
created automatically and a chown operation was performed for this user
immediately prior to a subsequent change of uid/gid values by puppet.
The result was that the services didn't start correctly.

This commit simply changes the ordering of the manifest, so that puppet
uses its default order to create the user before installing the package.

It also swaps the use of  require_package for ensure_package in-line
with current standards.

Bug: T255148
Change-Id: I28f87c37c3a3e8631857935b27af39878621e525