cache: use confd only if backend list is on etcd
The list of backends (ie: origin servers) used by varnish-fe is usually
taken from etcd. We use confd to render the list of pooled backends in a
VCL file, /etc/varnish/directors.frontend.vcl. However, we also allow
for the list of backends to be defined statically in the case of the
"single backend experiment", as well as in test deployments on Wikimedia
Cloud. Use confd only if the list of backends is to be taken from etcd.

Bug: T288106
Change-Id: I9b531fee73269cbf68cc0106f085088c6414ef45