mediabackup: Puppetize the media backup workers
Setup the configuration and paths needed to run the backups on each
backup "worker", the hosts where the backup logic runs and downloads
the necesary metadata and files and then send them to the final
storage location on the media backup storage hosts.

This requires access to multitude of hiera configuration endpoints,
some common, some per-datacenter, as well as access secrets for
the storage and the databases.

For now, we will hardcode just a backup of a single wiki. In the
future we will move some of the configuration to the database, due
to it being quite dynamic (e.g. list of wikis to backup)- keeping
it in sync automatically with production configuration.

Bug: T276442
Change-Id: I17695bbd272b27144534faf5606f9c8648096ab9