maintain-dbusers: delete users that are removed from ldap
In order to delete tool accounts we delete (mostly), pretend to delete
database accounts for users that have been removed from LDAP.

This could be generalized for PAWS, but that is not in scope.
Additionally, this first patch does not delete anything because we
want logs of what it thinks it should do first.

It calls for a followup patch to remove the "dryrun" argument.

Alternatively, this can easily check the user home dir or other
methods of determining it should remove things.

Bug: T285332
Change-Id: Ib86f995e7739538e92f9486a96e6255fcad03907