wikireplicas: add labswiki manually to s6 and refactor a bit
The "- all" value technically doesn't work anymore, so honestly,
that all needs to be stripped out of this. For now, this is much easier
to review.

Instead of a static list that adds a database to s7, now the config
can add arbitrary lists of databases to arbitrary replica sections. This
is prompted by the addition of labswiki to s6 with it not appearing in
https://noc.wikimedia.org/conf/dblists/s6.dblist

Bug: T287442
Change-Id: I1783ebaea732f8b94293a0860c3846d64d662519