profile::gitlab rsync fix rsync backup command
When doing rsync to a remote host, the module name in the rsyncd
config has to be used instead of the full path. If the full path
is used, rsync fails with @ERROR: Unknown module.
This change fixes a bug introduced in
Ia75a03d88e0dc2e0536a3a97055de12755b9cc27

Bug: T285867
Change-Id: I0abc7ecc2259df56e5ad076a34c1e4c18d4f2fbd