dbbackups: Fix wrongly configured target host for s7 (buster)
s7 on eqiad was being sent incorrectly to dbprov1001, failing
to prepare as it was a stretch host with mariadb 10.1 and the
source host was already in buster with mariadb 10.4.

Fix this to the right, intended host for preparing, dbprov1002.

This is a followup to 3ff5a11e288e73f450f645d374751a4b2d76491d

Bug: T288244
Change-Id: I4dc0a9e3d06513352453fe41299da851577a2d7a