deploy-mwdebug: introduce interactive mode
Given we want not only to be able to deploy mwdebug automatically but
also to make our own changes which might earn a more cautious approach,
an "interactive" mode is introduced for allowing human-driven deployments
using the script, that will thus make the automatic deployments stop for
the time that the deployment is running and only re-start once it's done.

Also allow changing the images to deploy.

Change-Id: I980d83c90d2783fa72ba9be8e23690314ae42fad