bacula: Change default retention to 60 days
We aim for a retention of 90 days. Sometimes jobs and files are
pruned too early, much before the volumes are pruned. Increase
it a bit to make sure files are available for longer.

Increase it to 60 days, which we are sure we can handle on
all pools, we may want to tune it later per type of client
(production vs Databases vs others), increasing it where
available.

This is possible due to the increase in available space.
We may want to monitor the current retention per job later.

This undos 5a54a521c29a, as it said "After we get more space,
we can proceed with more tuning"

Change-Id: I582731e9f37207959409dc4dcd2caa02634b6236