site/conftool: add mw1451 through mw1456 as apppservers, A1, A8
We currently have 54 appservers and 63 API servers and 17 jobrunners.

Not sure if canary API servers should be counted but either way
we should use these remaining 6 servers are appservers.

These are in newly used racks A1 and A8, contributing to lower the
percentage of servers in row D and pushing it under 40% as desired.

Bug: T279309
Change-Id: Icca1759c59b4d3487d2877103ca5c049f4e98b87