Nova-fullstack: improve DNS resolution tests
Previously we checked to see if the hostname resolved but didn't actually check
to see what it resolved to. Dig returns 0 on an NXDOMAIN so this test didn't actually
confirm anything; it moved on if there was no record or if there was the wrong record.

Now we check the results and throw an exception if dig returns a surprising IP; it also
times out if we get nxdomain for too long.

Bug: T288854
Change-Id: I48141f1657b699cdb23d6b8882d3ebe8bd53b001