backup: Update minio storage location to be /srv/objectstorage
In order to avoid a path called /srv/mediabackups/mediabackups

(first the mount point and second the container), rename the first
to objectstorage, also thinking that in the future it this object
storage could be used for things other than media backups.

Also reduce the batch size while initial run to 1000 rows/files.

Bug: T276442
Change-Id: Ia895dab57d381c44cf3a9f0158552e78151d9167