profile::gitlab load rsync::server only on passive GitLab
The rsync::server module is only needed on the passive instance.
In Ia75a03d88e0dc2e0536a3a97055de12755b9cc27 rsync between the
active and passive GitLab was introduced. Currently the
rsync::server module is loaded on both instances. On the active
instance, this leads to change on every puppet run because the class
is loaded and stopped immediately. See:
https://puppetboard.wikimedia.org/node/gitlab1001.wikimedia.org

Bug: T285867
Change-Id: Ib850b430ed1b05f328f2de2c35273ab02df0e1bc