hieradata: remove 9 redis shards
Following I78ba6bd3e74fb0e16f7de670e8d6950166425afd
we are removing the rest 9 shards, leaving a total of
8 shards. We are removing shards hosted on odd numbered
servers, with the exception of mc1028 as its shard 
counterpart was an odd server. 

Lastly, shards are not renamed as nutcracker uses a 
shard's name as a hashing key. 

Bug: T280582
Change-Id: I21a00fea0839db3a23a62869827ad82de8902753