url_downloader: Don't cache ICMP database
Why:
squid has an ICMP RTT database that it can share with other squid
neighbors in order to speedup cache lookups by clients. In
url_downloaders we don't have a cache to start with and we have ICP (the
protocol for sharing with neighbors) disabled. There is thus, no reason
to persist the database (called netdb). However it is persisted by
default and we want to disable that.

What:
Pass netdb_filename none. This will make frequently accessed 3rd party
sites just a tad a little bit slower (order of ms) after a squid restart
However, since url downloaders are not used directly by end-users by
only services, that should be totally ok.

Bug: T286525
Change-Id: I910afc79f29a5c1325ad21ad8d894a605aa2581d