openstack: stop paging for systemd alone
It makes sense to intentionally set alerts on things we want pages for.
However, many things are email-only alerts for backups and such that
we really don't need to get out of bed for. With this set, systemd will
not page us for any failure that is not specifically set to page.

We've talked about this many times, but let's discuss on this patch.

We can remove it if we can change profile::base enough to allow us to
just get emails or IRC alerts for systemd generic issues.

Bug: T280493
Change-Id: I91ad7e23fe52d1cb373733b42b3ebe7443df1b84