mediabackup: Add new conf and identity for read only recovery
When recovering data from backups, we want to make sure we don't
alter the existing ones. Create a new identity (access key and
secret key pair) that will have read-only priviliges.

In fact, we won't be using the built-ing readonly policy for this
account, but a newly created policy, readandlist, as we want to
be able to list the bucket files in addition to GET them.
The custom policy is also added to the mediabackup config files.

Set up a new file with the configuration details for this. This
will be useful also for checking and statistics.

This will require the addition of the account and key, per
datacenter on private hiera.

Bug: T276442
Change-Id: Ieb6edf15b0833559a5ccbe32e39515cb60f8668b