update quarry systemd and branch
currently quarry will pull from the "master" branch, for
this update we want to be on the buster branch and it seems
we will remain there. Additionally newer celery will need
a different order to its startup command.

This change does bring up some deprecated code warnings for
celery, so long as it works we will want a ticket to update
those before we update celery again.

This will likely break stretch if merged, we should be able
to do the updates manually on the buster systems until we
are ready to merge this. Do not merge until then.

Change-Id: Ic1032365d91333daf13ca4dbd328140ecedc5d6c