dbbackups: Move s4 backup generation from db1130 to db1150/dbprov1003
By the time this patch is merged, all eqiad mariadb mw instances
for s4 (commonswiki and testcommonswiki) will have been migrated
to Debian Buster/MariaDB 10.4. Start generating 10.4 tarballs
for snapshot backups on eqiad. For that, we will have to switch
backups from db1139 (still in stretch) to db1150 (already upgraded
to buster), as well as prepare them on a buster dbprov host
(dbprov1003 in this case).

Bug: T288803
Change-Id: I9ab032eca8dfae7a49effd9218b795b867fab49d