lldp fact: updated lldp parent fact to fall back to routers
Currently the LLDP fact is unable to correctly determine the lddp parent
on ganeti hosts.  this is because the primary interface is not the
physical interface connected to the switch.

This CR updates the lldp fact to indicate if the switch is advertising
the router capability.  This in formation is then used when determining
the lldp_parent.  if no parent is found using the current method then we
will use the first lldp neighbour that claims to be a router

Bug: T289679
Change-Id: I4b5eab82822f98f2a5b831464757bfafa53fffec