P:base: move production specific code to their own profile
Currently all host in cloud and production environments have an entry
point of profile::standard which in turn loads profile::base.

profile::base includes many modules which are only really relevant to
the production environment which are currently instantiated on the cloud
environment just because of the fact that both environments use the same
entry point.  This CR introduces a new profile::base::production which
will be used to define base profiles and classes which only belong to
the production environment.

This will reduce the amount of redundant code running on cloud instances
and further, hopefully, reduce the chance of changes to production code
affecting/breaking cloud instances.

With this in mind the following modules and profiles have been identified
as productions specific and as such will be moved to the new
profile::base::production class and thus will not be instantiated
on cloud instances.

    * profile::admin
    * profile::pki::client
    * profile::contacts
    * profile::base::netbase
    * profile::logoutd
    * profile::cumin::target
    * profile::debmonitor::client
    * base::phaste
    * base::screenconfig
    * toil::acct_handle_wtmp_not_rotated
    * profile::monitoring
    * profile::emacs
    * interface::add_ip6_mapped

The result of this is that post change profile::base will only contain
classes which are really shared between production and cloud
environments

Bug: T289661
Hosts: P:standard
Change-Id: I02a8ec7d102da1058daaca8b38237114ccb1bde9
Signed-off-by: David Caro <dcaro@wikimedia.org>