Revert "nova_fullstack: try to get the puppet state from a couple places"
Apparently a working puppet run changes the state dir from /var/cache to /var/lib.  So
we /want/ to ignore the file in /var/cache in favor of /var/lib because having the file
in /var/lib suggests that things have worked right.

1. First (good) puppet run changes the state dir to /var/lib but doesn't write anything there
2. Second (good) puppet run records the last_puppet_run file in /var/lib

Since our firstboot process runs puppet twice in a row, we can depend on finding that file
in /var/lib if all is well.

This reverts commit 67678034fbf05607ff7c64f7417cbd66783d57f5.

Bug: T289663
Change-Id: I40a3d6a4c7d987399adeb8804cd0b0f7af7431e1