lldp fact: add new parent key to lldp
add a new implementation of lldp_parent to the lldp fact.  Once this fact
is working we will update lldp_parent to alias it.

The new fact first tries to set the parent based on the following
order:
 * the neighbour advertised on the primary interface
 * the first neighbour to advertise the Router capability that does not
   have a Linux description

The vast majority of machines set their parent correctly based on the
primary interface, however the ganeti hosts don't receive any lddp
advertisment over the interface listed as primary as such this fails.
In this CR we introduce parsing of the router attribute and set the
parent based on this value.  We explicitly exclude advertisments with
the word 'Linux' in the description (Linux servers advertise
`uname -a` in their desc) we only see this issue on baremetal and as
such expect to get a switch/router and not a linux host.

Bug: T289679
Change-Id: I8af261cd55c8a3a5c94bd23eb771639b10bd3338