create a generic class to clean the puppet client bucket
A systemd timer to empty the puppet client bucket data which can grow
large.

Create a generic version of the existing timer in profile::base::labs
to be used in production.

Turn it into a profile that can be included in roles where it is desired.

Switch the default to 'true' instead of opting in in Hiera.

Bug: T165885
Change-Id: I8304f6aeb1c71c395284e646fa9e004f278f7e7d