admin: Update approver of analytics groups
This change is to add Olja as the person documented as responsible to
aprove changes of membership to analytics groups, as the Data
Engineering Director. That is, sign off who gets and doesn't get access
to certain private data available on WMF servers.

This was already happening in practice, but documentation on
the file was left unchanged.

Change-Id: I504a14290405b99219bfe585c545284250d512e8