deployment: Use rsync::quickdatacopy, enable encryption
Instead of manually setting up rsync modules and timers, use
rsync::quickdatacopy to do that now that it supports setting
custom intervals and having multiple destination hosts.

The primary motivation for this change is to enable encryption
as some of these rsyncs will be cross-DC. The rsync is now
wrapped in stunnel for that.

Bug: T289857
Co-Authored-by: claime <cgoubert@mediawiki.org>
Hosts: O:deployment_server
Hosts: deploy2002.codfw.wmnet,deploy1002.eqiad.wmnet
Change-Id: I753551ca569785045b9e57fdffee9453c7acbcbb