admin: drop dependencies between adminuser and admingroup
The admin::user resources creates a user type that uses a gid and we need
to ensure that gid exists before we create the user.  however the user
type has an autorequire on groups as such we don't need this explicit
require.  the important thing is to ensure we have a require between the
command that updates group membership and all users

Once merged we should test this with a reimage

Bug: T263578
Hosts: C:admin
Change-Id: I4cf12b4655bd42bf7aaccfe09cd3b8548d2124dc