admin: create a group to run the wmf-auto-reimage commands
Currently cookbooks,
and these commands which will be converted to cookbooks,
require root access (though there is a goal
to change that in the future).

Arnold does not currently have global root (he does have gitlab
root).

I am working with Arnold to show him how to install an OS on some
new servers, using the cookbook. As a next step it would be nice
if he could run the commands himself on a cumin host.

So before we get into a discussion about global root and/or
letting people run this specific cookbook without it, I suggest
a new admin group as a stop gap.

It's not actually hard to create and we can see it as temporary
and remove it again if it's not needed anymore.

This is supposed to cover both wmf-auto-reimage and wmf-auto-reimage-host
with any host name as parameter.

Hosts: cumin2002.codfw.wmnet
Change-Id: I7ae14d72c23927fbe08ebe5221fe26e9a2259736