dbbackups: Migrate s4 generation from db2097 (stretch) to db2139 (buster)
Start generating MariaDB 10.4 backups as, at the time of the merge,
the primary s4 instance for codfw has been (or soon will be) migrated
to a Debian Buster/MariaDB 10.4 instance. We will stop taking backups
at db2097:3314 and start doing it at db2139:3314.

We also have to prepare snapshots on a buster host with xtrabackup
compiled against 10.4, so we will use dbprov2003 for s4.

Bug: T288803
Change-Id: I98e5a60a6f7231ecc4bb0342e44aac4a1a8fb41e