facter networking: filter out cali/tap interfaces
Having large facts on hosts can cause delays when submitting to
puppetdb. Try and trim the networking fact to remove noise and the, I
believe, unused cali interface from the networking.interfaces fact.
Also filter out the lo:LVS interface which has a list of all the
configured service IP addresses

Similar to the cali interfaces we also filter tap interfaces which are
used created by ganeti/openstack

Ideally we would be able to filter theses via puppetdb's
facts-blacklist, however we are only able to filter out the the entire
fact or nothing at all. Since we query for the networking fact directly
from puppetdb in e.g. netbox reports, removing the entire networking
fact from puppetdb is not an option

As such this CR instead filters the fact in facter and thus means
theses values will be completely removed and not available to use in
puppet code.

We can still query for $facts['networking']['interfaces'] but it will
nolonger have interfaces named 'cali.......'

Bug: T265904
Change-Id: I5fefea4b339f336b264e7e5873bdd5f2aaaea2ec