airflow: Compress scheduler logs
The airflow scheduler logs are limited to 2 weeks and still take
multiple GB of space. In a quick test these logs compress ~20x,
meaning our 5GB of scheduler logs should turn into a few hundred
MB. This has to be limited to the scheduler as the task logs are
viewable in the airflow ui and need to stay available in their
original form.

Change-Id: Ib34b08bdca5aa01e65a8b502d05a47e73a9f84f2