Add image suggestion api to lookup table
The Image Suggestion API is a proof-of-concept API
that is not yet on K8s, but would like to have some
logging to see how it would work for an application
not yet on kubernetes. The API is setup as a systemd
service named image-suggestion-api, and outputs its logs
to stdout. This was my best guess at what should be in this
file!

Bug: T288132
Change-Id: If4ad7e800180b7d11a39f38dd7bbbc41d34e3d4b