Run GrowthExperiments fixLinkRecommendationData --dry-run every day
The database table growthexperiments_link_recommendations and the
recommendation.link weighted tag in the CirrusSearch index need to
be in sync; this has been a source of problems in the past.

Run fixLinkRecommendationData in dry-run mode daily to count the
number of search index and database entries which do not have a
matching record, and report the results to statsd.

Tooks about 2 minutes on the current set of Growth wikis.

Bug: T283868
Change-Id: I2050e052c7ce5c7b099bd9aa37e96801d145dbcf