rsyslog: expand output lookup table docs
The $.log_outputs default value for $programname in case of no match is
"local", meaning that logs are saved locally on disk by rsyslog based on
facility and priority if there is no specific override in the lookup
table.

Expand the comment in 95-output-local.conf a bit to clarify how the
lookup works.

Bug: T206454
Bug: T206633
Change-Id: I12cd9a9af1c8b708fe3cb8ee86492dc69e2f6038