wdqs: temp mitigation => restart hourly w random
We only want to perform this mitigation on codfw wdqs, so only restart
if the hostname starts with `wdqs2`

We wanted to add randomization with systemd's RandomizedDelaySec, but
systemd::timer::job doesn't expose that to us so it would add too much
complexity. So instead we hack it in the bash script. This will make it
restart between 55 and 65 minutes, for an average of once per hour

Hosts: wdqs2001.codfw.wmnet, wdqs1003.eqiad.wmnet
Bug: T290330
Change-Id: I43c057d9d7505a0fb716476c01b84d535af6a271