conftool-data: add worker nodes to ml_serve
The ml_serve alias is currently only showing kubemaster nodes,
and we need also the "kubesvc" tags to be able to create
a DNS discovery-based LVS service in front of them
(that will be inference.discovery.wmnet).

We are keeping the same names/tags as the "kubernetes"
as ml_serve is another kubernetes cluster.

Bug: T289835
Change-Id: Ia80e461d28def98c861242a50ea3ca3ef83a2aaa