Add a prometheus scrape target for the aqs_next role
This is the new cluster for AQS, which is currently not being scraped by
prometheus. We are going to add it whilst the hosts are in the aqs_next
role, so that we can monitor performance during the migration. Later we
will remove this and rename the new cluster to aqs.

Bug: T249755
Change-Id: Ia9c18c071ff6da4fc4f85354847a79f84ce222df