cloud/devtools: set docker::registry to localhost
puppet is broken since a while on deployment servers
in cloud VPS.

This is because docker::registry is expected to be a FQDN
and in hieradata/cloud/common it is set to just ''.

Set it randomly to localhost for just this project to fix the
puppet run there. Not sure what else to use.

Change-Id: Ib84939c210e07ddd741bab03f582f0d7129957b6