P:base: drop broad dependency
Broad dependencies have been shown to causes issues with puppetdb in the
past.  As such lets drop the automatic require and instead contain
profile::base::certificates.  The initial issue which promted this
dependency related to the puppet ca and reimaging a host, as such
contianing should ensure update-ca-certificates is run very early.

Bug: T244477
Bug: T263578
Change-Id: I84b6056d4dbbc14f083515f6de46c1a93030a7c6