puppetmaster: drop log messages from logstash reporter
It seems that this report never got used in production as we didn't want
to expose diff messages (which may include private data).  Further i
this file doesn't seem to be used in cloud (check with cloud).

As such i propose simply dropping the log messages from event data.  The
other data (ill post an example in the comments) is still very useful.
Further i'm not sure the log messages would be that useful in the first
place, it is an array of free text lines so not great to search, but i
could be missing something and we could add it back when we have acl for
logstash

example output:
  https://phabricator.wikimedia.org/P17608

Bug: T222826
Depends-On: Ib7f1607497df6d006bf67c91976930e825f37823
Change-Id: Ifec43bd0fe9dd7ee8e1f4283e8f2c8f514d3d1b1